% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fce_bullshark_doc.R
\docType{data}
\name{fce_bullshark}
\alias{fce_bullshark}
\title{Bull shark catches, water temperatures, salinities, and dissolved oxygen levels in the Shark River Slough}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 255 rows and 11 columns.
}
\source{
{Heithaus, M. and P. Matich. 2014. Bull shark catches, water temperatures, salinities, and dissolved oxygen levels in the Shark River Slough, Everglades National Park (FCE) , from May 2005 to May 2009 ver 3. Environmental Data Initiative. https://doi.org/10.6073/pasta/04a8792fed9ceed4237bd3273a97e8f8 (Accessed 2021-02-20).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-fce.1167.3}
}
\usage{
fce_bullshark
}
\description{
This dataset provides information on the catches of bull sharks in the Shark River Slough in relation to physical factors including dissolved oxygen, water temperature, salinity, and distance upstream.
}
\keyword{datasets}
